Ensured that Distributed object's are constructed using privileges equivalent to CodeSource(null, null), very minimal to avoid the ability to prevent remote processes from creating privileged objects. See SerialReflectionFactory.
Updated UriTest

Ensured visibility of test fields by changing to final or volatile.

Changed LookupDropProxyTaskRace test so it isn't run by default, it must be run manually.


git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/skunk/qa_refactor/trunk@1479174 13f79535-47bb-0310-9956-ffa450edef68

+++ src/net/jini/lookup/ServiceDiscoveryManager.java
@@ -3877,7 +3877,7 @@ public class ServiceDiscoveryManager {
                                           (COMPONENT_NAME,
                                            "discardWait",
                                            long.class,
-                                           init.discardWait)).longValue();
+                                           init.discardWait));
         /* Discovery manager */
         init.discMgr = discoveryMgr;
 	if(init.discMgr == null) {
